Update CSS selector in scraper.py
In September 2021, the specific CSS selector for the image thumbnails on Google Images changed. I inspected the Google Images website and found the new CSS selector by following the steps mentioned in your blog. It was really helpful. To ensure that others don't face any issues, I am submitting a pull request with the modification of the CSS selector only.